# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Slogr Controller API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/add-profile": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a profile",
                "description": "Create a new profile",
                "operationId": "22f7c233776c3d20259c5d9390549bb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Profile name",
                                        "type": "string",
                                        "example": "Profile 1"
                                    },
                                    "p_size": {
                                        "description": "Packet size",
                                        "type": "integer",
                                        "example": "1500"
                                    },
                                    "count": {
                                        "description": "Count",
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "n_packets": {
                                        "description": "Number of packets",
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "p_interval": {
                                        "description": "Packet interval",
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "w_time": {
                                        "description": "Wait time",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "dscp": {
                                        "description": "DSCP",
                                        "type": "integer",
                                        "example": "46"
                                    },
                                    "max_loss": {
                                        "description": "Maximum loss",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "max_down": {
                                        "description": "Maximum downlink time",
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "max_up": {
                                        "description": "Maximum uplink time",
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "max_jitter": {
                                        "description": "Maximum jitter",
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "max_rtt": {
                                        "description": "Maximum RTT",
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "edit": {
                                        "description": "Edit flag",
                                        "type": "boolean",
                                        "example": "false"
                                    },
                                    "aid": {
                                        "description": "Profile ID (required if edit flag is true)",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "profile": {}
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "300": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete-profile": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a profile",
                "description": "Delete a profile by ID",
                "operationId": "3c5c7640dc595d01f0d6c7ea9942b027",
                "requestBody": {
                    "description": "Profile ID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "delete": {
                                        "description": "ID of the profile to be deleted",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the deletion operation",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/push/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Push profile to clients",
                "description": "Push the specified profile to clients",
                "operationId": "bdc16da4f078c742b589720095e4eef3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "security": {
                "passport": {
                    "type": "oauth2",
                    "description": "Laravel passport oauth2 security.",
                    "in": "header",
                    "scheme": "https",
                    "flows": {
                        "password": {
                            "authorizationUrl": "localhost/slogr/oauth/authorize",
                            "tokenUrl": "localhost/slogr/oauth/token",
                            "refreshUrl": "localhost/slogr/token/refresh",
                            "scopes": []
                        }
                    }
                }
            }
        }
    }
}